feat: RoomFactory findByPosition 기능 구현
