+++ src/main/resources/templates/contents/chessBoard.html
@@ -102,6 +102,5 @@
             findById("{{position}}", "{{symbol}}");
         {{/chessPiece}}
     }
-
 </script>
 </html>
